[refactor] Refactor Test code 테스트코드 반영 related to #36
