+++ search/src/test/java/com/bjcareer/search/service/StockServiceTest.java
@@ -15,7 +15,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import com.bjcareer.search.application.stock.StockService;
 import com.bjcareer.search.domain.entity.Stock;
 import com.bjcareer.search.out.crawling.naver.CrawlingNaverFinance;
-import com.bjcareer.search.out.repository.stock.StockRepository;
+import com.bjcareer.search.out.persistence.repository.stock.StockRepository;
 import com.bjcareer.search.application.exceptions.InvalidStockInformation;
 
 @ExtendWith(MockitoExtension.class)
